Added new output formats, delta based
Added prototype output formats which only output the changed sid registers per tick, by using a format which consists of:

(1 byte of #of regs to update) followed by (reg,val) pairs of values equal to the number of regs to update.
Note that the HI/LO bytes of the timing(a count value in us) are treated as registers too, with numbers of 25 and 26 in the format.